+++ rosdev/gen/docker/gdbinit.py
@@ -2,11 +2,12 @@ from atools import memoize
 from dataclasses import dataclass
 from logging import getLogger
 from pathlib import Path
+from textwrap import dedent
 
-from rosdev.gen.docker.container import GenDockerContainer
 from rosdev.gen.home import GenHome
 from rosdev.gen.host import GenHost
-from rosdev.gen.rosdev.workspace import GenRosdevWorkspace
+from rosdev.gen.rosdev.home import GenRosdevHome
+from rosdev.gen.src import GenSrc
 from rosdev.util.handler import Handler
 from rosdev.util.options import Options
 
@@ -17,46 +18,48 @@ log = getLogger(__name__)
 @dataclass(frozen=True)
 class GenDockerGdbinit(Handler):
 
-    @classmethod
+    @staticmethod
     @memoize
-    async def get_path(cls, options: Options) -> Path:
-        path = await GenRosdevWorkspace.get_path(options) / 'gdbinit'
-        
-        log.debug(f'{cls.__name__} {path = }')
-        
-        return path
+    async def get_container_path(options: Options) -> Path:
+        container_path = await GenHome.get_path(options) / '.gdbinit'
 
-    @classmethod
+        log.debug(f'{GenDockerGdbinit.__name__} {container_path = }')
+
+        return container_path
+
+    @staticmethod
     @memoize
-    async def get_symlink_path(cls, options: Options) -> Path:
-        symlink_path = await GenHome.get_path(options) / '.gdbinit'
+    async def get_home_path(options: Options) -> Path:
+        home_path = await GenRosdevHome.get_path(options) / 'gdbinit'
+        
+        log.debug(f'{GenDockerGdbinit.__name__} {home_path = }')
+        
+        return home_path
 
-        log.debug(f'{cls.__name__} {symlink_path = }')
-
-        return symlink_path
+    @staticmethod
+    async def get_text(options: Options) -> str:
+        text = dedent(f'''
+                set directories {
+                    ":".join([
+                        f"{path.parent}"
+                        for path in (await GenSrc.get_path(options)).rglob("CMakeLists.txt")
+                    ])
+                }
+        ''').lstrip()
+        
+        log.debug(f'{GenDockerGdbinit.__name__} {text}')
+        
+        return text
 
     @classmethod
     async def main(cls, options: Options) -> None:
+        if (await cls.get_home_path(options)).exists():
+            return
+
         log.info(f'Creating docker_gdbinit')
         GenHost.write_text(
-            # FIXME these are two common paths from ci.ro2.org builds. Find a way to
-            #  programmatically find the paths, probably through jenkins.
-            data=(
-                f'set directories {options.src_symlink_path / "ros2"}'
-                f'set substitute-path '
-                f'/home/jenkins-agent/workspace/'
-                f'packaging_{options.operating_system}/ws/src/ '
-                f'{options.src_symlink_path}\n'
-                f'set substitute-path '
-                f'/home/rosbuild/ci_scripts/ws/src/ '
-                f'{options.src_symlink_path}\n'
-            ),
+            data=await cls.get_text(options),
             options=options,
-            path=options.docker_gdbinit_path,
+            path=await cls.get_home_path(options),
         )
-        await GenDockerContainer.execute(
-            commmand=f'ln -s {await cls.get_path(options)} {await cls.get_symlink_path(options)}',
-            options=options,
-        )
-
         log.info(f'Created docker_gdbinit')
